\name{check.for.library}
\alias{check.for.library}
\title{Check for library}
\description{Check if library is available}
\usage{
  check.for.library(lib.name = 'valection');
}
\arguments{
  \item{lib.name}{A string of the library name. Defaults to 'valection'.}
}
\details{Used internally to check if executable is available in PATH.}
\value{A logical indicating presence of library}
\author{Dorota Sendorek}
\keyword{internal}
