\name{handle.missing.library}
\alias{handle.missing.library}
\title{Handle missing library}
\description{Output error message for missing library}
\usage{
  handle.missing.library(lib.name = 'valection');
}
\arguments{
  \item{lib.name}{A string of the library name. Defaults to 'valection'.}
}
\details{Used internally to output an error message for missing library.}
\author{Dorota Sendorek}
\keyword{internal}
