\name{quoted}
\alias{quoted}
\title{Enquote a string}
\description{Add single quotation marks around a string}
\usage{
  quoted(string);
}
\arguments{
  \item{string}{A string}
}
\details{Used internally to create bash commands.}
\value{A string}
\author{Chris Cooper and Dorota Sendorek}
\keyword{internal}
